Add basic site structure, about, login/registration
I think I did two hours of work on this on a plane several months ago, and am picking it up again now. The registration process is actually the core of the actor experience, so that needs to get significantly edited next. And then we can work on theater account functionality.